Madara module and mangarockteam test
